+++ examples/classes.go
@@ -12,57 +12,93 @@ type Speaker interface {
 }
 
 //line examples/classes.zn:5
-type Animal struct {
+type AnimalImpl struct {
 	Name string
 }
 
-func NewAnimal(name string) *Animal {
-	obj := &Animal{}
+func (a *AnimalImpl) SetName(v string) { a.Name = v }
+
+type Animal interface {
+	SetName(string)
+	GetName() string
+}
+
+var _ Animal = (*AnimalImpl)(nil)
+
+func NewAnimal(name string) *AnimalImpl {
+	obj := &AnimalImpl{}
 //line examples/classes.zn:9
 	obj.Name = name
 	return obj
 }
 
-func (a *Animal) GetName() string {
+func (a *AnimalImpl) GetName() string {
 //line examples/classes.zn:13
 	return a.Name
 }
 
 //line examples/classes.zn:17
-type Dog struct {
-	Animal
+type DogImpl struct {
+	AnimalImpl
 	Breed string
 }
 
-var _ Speaker = (*Dog)(nil)
+func (d *DogImpl) GetBreed() string  { return d.Breed }
+func (d *DogImpl) SetBreed(v string) { d.Breed = v }
 
-func NewDog(name string, breed string) *Dog {
-	obj := &Dog{
-		Animal: *NewAnimal(name),
+type Dog interface {
+	Animal
+	Speaker
+	GetBreed() string
+	SetBreed(string)
+	Speak() string
+	Describe() string
+}
+
+var _ Dog = (*DogImpl)(nil)
+var _ Speaker = (*DogImpl)(nil)
+
+func NewDog(name string, breed string) *DogImpl {
+	obj := &DogImpl{
+		AnimalImpl: *NewAnimal(name),
 	}
 //line examples/classes.zn:22
 	obj.Breed = breed
 	return obj
 }
 
-func (d *Dog) Speak() string {
+func (d *DogImpl) Speak() string {
 //line examples/classes.zn:26
 	return "Woof!"
 }
 
-func (d *Dog) Describe() string {
+func (d *DogImpl) Describe() string {
 //line examples/classes.zn:30
 	return d.GetName()
 }
 
-//line examples/classes.zn:34
-func main() {
 //line examples/classes.zn:35
-	d := NewDog("Rex", "Labrador")
+func greetAnimal(a Animal) {
 //line examples/classes.zn:36
+	fmt.Println(fmt.Sprintf("Hello, %v!", a.GetName()))
+}
+
+//line examples/classes.zn:39
+func makeSpeak(s Speaker) {
+//line examples/classes.zn:40
+	fmt.Println(s.Speak())
+}
+
+//line examples/classes.zn:43
+func main() {
+//line examples/classes.zn:44
+	d := NewDog("Rex", "Labrador")
+//line examples/classes.zn:45
 	fmt.Println(d.Speak())
-//line examples/classes.zn:37
+//line examples/classes.zn:46
 	fmt.Println(d.Describe())
-//line examples/classes.zn:38
-	fmt.Println(d.GetName())
+//line examples/classes.zn:47
+	greetAnimal(d)
+//line examples/classes.zn:48
+	makeSpeak(d)
 }