+++ examples/myapp/main.go
@@ -1,5 +1,3 @@
-//go:build ignore
-
 package main
 
 import (
@@ -8,11 +6,18 @@ import (
 	"myapp/utils"
 )
 
+//line main.zn:4
 func main() {
+//line main.zn:5
 	sum := utils.Add(10, 32)
+//line main.zn:6
 	product := utils.Multiply(6, 7)
+//line main.zn:7
 	fmt.Println(sum)
+//line main.zn:8
 	fmt.Println(product)
-	dog := models.NewDog("Rex")
+//line main.zn:10
+	var dog interface{} = models.NewDog("Rex")
+//line main.zn:11
 	fmt.Println(dog)
 }